Refactor: user domain 추가[박한솔]
