--- basilium-server/src/main/java/basilium/basiliumserver/domain/SuperUser.java
@@ -1,11 +1,0 @@
-package basilium.basiliumserver.domain;
-
-import jakarta.persistence.Entity;
-
-@Entity
-public class SuperUser extends User{
-    public SuperUser() {
-
-        super();
-    }
-}
