+++ basilium-server/src/main/java/basilium/basiliumserver/domain/user/NormalUser.java
@@ -1,5 +1,6 @@
-package basilium.basiliumserver.domain;
+package basilium.basiliumserver.domain.user;
 
+import basilium.basiliumserver.domain.user.User;
 import jakarta.persistence.Entity;
 import lombok.Getter;
 import lombok.Setter;
@@ -7,7 +8,7 @@ import lombok.Setter;
 @Entity
 @Getter
 @Setter
-public class NormalUser extends User{
+public class NormalUser extends User {
     private String name;
     private Long age;
     private String address;
