+++ basilium-server/src/main/java/basilium/basiliumserver/domain/user/SuperUser.java
@@ -1,0 +1,12 @@
+package basilium.basiliumserver.domain.user;
+
+import basilium.basiliumserver.domain.user.User;
+import jakarta.persistence.Entity;
+
+@Entity
+public class SuperUser extends User {
+    public SuperUser() {
+
+        super();
+    }
+}
